refactor: split out systmes
